pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyglplot"
version = "0.1.3"
authors = [
  { name="Danial Chitnis", email="d.chitnis@ed.ac.uk" },
]
description = "A Python package for plotting."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "numpy",
  "glfw",
  "pyopengl",
]

[project.urls]
"Homepage" = "https://github.com/danchitnis/pyglplot"
"Bug Tracker" = "https://github.com/danchitnis/pyglplot/issues"

[tool.hatch.build]
exclude = [
  "tests",
  "test/*",
  "test/**",
  ".github",
  "docs",
  "temp",
  ".vscode"
]